Reorganise classes
